refactor(exception): introduce ErrorInfoCapable interface for unified error handling
- Add ErrorInfoCapable interface to standardize error information retrieval
- Implement ErrorInfoCapable in CommandResultException- Update ErrorConverter to handle ErrorInfoCapable objects
- Remove redundant error handling logic in RequestExceptionHandler
- Add unit tests for new error handling approach
