+++ wow-core/src/main/kotlin/me/ahoo/wow/exception/ErrorConverter.kt
@@ -14,6 +14,8 @@
 package me.ahoo.wow.exception
 
 import me.ahoo.wow.api.exception.ErrorInfo
+import me.ahoo.wow.api.exception.ErrorInfo.Companion.materialize
+import me.ahoo.wow.api.exception.ErrorInfoCapable
 import java.lang.reflect.ParameterizedType
 import java.util.concurrent.TimeoutException
 
@@ -36,6 +38,12 @@ abstract class AbstractErrorConverterFactory<E : Throwable> : ErrorConverterFact
 
 object DefaultErrorConverter : ErrorConverter<Throwable> {
     override fun convert(error: Throwable): ErrorInfo {
+        if (error is ErrorInfoCapable) {
+            return error.errorInfo.materialize()
+        }
+        if (error is ErrorInfo) {
+            return error.materialize()
+        }
         val errorCode = when (error) {
             is IllegalArgumentException -> ErrorCodes.ILLEGAL_ARGUMENT
             is IllegalStateException -> ErrorCodes.ILLEGAL_STATE
