+++ wow-core/src/main/kotlin/me/ahoo/wow/exception/ErrorConverterRegistrar.kt
@@ -16,7 +16,6 @@ package me.ahoo.wow.exception
 import io.github.oshai.kotlinlogging.KotlinLogging
 import me.ahoo.wow.annotation.sortedByOrder
 import me.ahoo.wow.api.exception.ErrorInfo
-import me.ahoo.wow.api.exception.ErrorInfo.Companion.materialize
 import java.util.*
 import java.util.concurrent.ConcurrentHashMap
 
@@ -60,9 +59,6 @@ object ErrorConverterRegistrar {
 }
 
 fun Throwable.toErrorInfo(): ErrorInfo {
-    if (this is ErrorInfo) {
-        return this.materialize()
-    }
     val errorConverter = ErrorConverterRegistrar.get(this.javaClass) ?: DefaultErrorConverter
     return errorConverter.convert(this)
 }
