+++ wow-webflux/src/main/kotlin/me/ahoo/wow/webflux/exception/RequestExceptionHandler.kt
@@ -14,7 +14,6 @@
 package me.ahoo.wow.webflux.exception
 
 import io.github.oshai.kotlinlogging.KotlinLogging
-import me.ahoo.wow.command.CommandResultException
 import me.ahoo.wow.exception.toErrorInfo
 import me.ahoo.wow.webflux.route.toServerResponse
 import org.springframework.web.reactive.function.server.ServerRequest
@@ -34,9 +33,6 @@ object DefaultRequestExceptionHandler : RequestExceptionHandler {
     override fun handle(request: ServerRequest, throwable: Throwable): Mono<ServerResponse> {
         log.warn(throwable) {
             request.formatRequest()
-        }
-        if (throwable is CommandResultException) {
-            return throwable.commandResult.toServerResponse()
         }
         return throwable.toErrorInfo().toServerResponse()
     }
